cli,cliccl: move mt commands to cliccl
Part of #91714
Epic: none

Release note: None
