+++ pkg/ccl/cliccl/BUILD.bazel
@@ -5,9 +5,14 @@ go_library(
     name = "cliccl",
     srcs = [
         "cliccl.go",
+        "context.go",
         "debug.go",
         "demo.go",
         "ear.go",
+        "flags.go",
+        "mt.go",
+        "mt_proxy.go",
+        "mt_test_directory.go",
         "start.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/cliccl",
@@ -16,21 +21,27 @@ go_library(
         "//pkg/base",
         "//pkg/ccl/baseccl",
         "//pkg/ccl/cliccl/cliflagsccl",
+        "//pkg/ccl/sqlproxyccl",
+        "//pkg/ccl/sqlproxyccl/tenantdirsvr",
         "//pkg/ccl/storageccl/engineccl/enginepbccl",
         "//pkg/ccl/utilccl",
         "//pkg/ccl/workloadccl/cliccl",
         "//pkg/cli",
         "//pkg/cli/clierrorplus",
         "//pkg/cli/cliflagcfg",
+        "//pkg/cli/cliflags",
         "//pkg/cli/democluster",
         "//pkg/storage",
         "//pkg/storage/enginepb",
+        "//pkg/util/log",
+        "//pkg/util/log/severity",
         "//pkg/util/protoutil",
         "//pkg/util/stop",
         "//pkg/util/timeutil",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_errors//oserror",
         "@com_github_cockroachdb_pebble//vfs",
+        "@com_github_cockroachdb_redact//:redact",
         "@com_github_spf13_cobra//:cobra",
     ],
 )
